requirements.txt
ws-api>=0.26.0
keyring>=24.0.0
pyyaml>=6.0

